requirements.txt
asgiref==3.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
idna==3.7
Markdown==3.5.1
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.2
